moving and remapping cohort views and templates
