+++ littleTalkApp/views_modules/profile.py
@@ -5,7 +5,7 @@ from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 
-from littleTalkApp.forms import CohortForm, LearnerForm
+from littleTalkApp.forms import LearnerForm
 from littleTalkApp.models import Cohort, Learner, LogEntry, Role
 
 
@@ -192,101 +192,3 @@ def confirm_delete_learner(request, learner_uuid):
         )
 
     return render(request, "profile/confirm_delete_learner.html", {"learner": learner})
-
-
-@login_required
-def cohort_list(request):
-    school = request.user.profile.get_current_school(request)
-    if not (
-        request.user.profile.is_admin_for_school(school)
-        or request.user.profile.is_manager_for_school(school)
-    ):
-        return redirect("profile")
-
-    cohorts = Cohort.objects.filter(school=school).order_by("name")
-    return render(
-        request,
-        "profile/cohorts/cohort_list.html",
-        {
-            "cohorts": cohorts,
-            "can_edit_cohorts": request.user.profile.is_admin_for_school(school)
-            or request.user.profile.is_manager_for_school(school),
-        },
-    )
-
-
-@login_required
-def cohort_create(request):
-    school = request.user.profile.get_current_school(request)
-    if not (
-        request.user.profile.is_admin_for_school(school)
-        or request.user.profile.is_manager_for_school(school)
-    ):
-        return redirect("profile")
-
-    if request.method == "POST":
-        form = CohortForm(request.POST)
-        if form.is_valid():
-            cohort = form.save(commit=False)
-            cohort.school = school
-            cohort.save()
-            return redirect("cohort_list")
-    else:
-        form = CohortForm()
-
-    return render(
-        request, "profile/cohorts/cohort_form.html", {"form": form, "is_editing": False}
-    )
-
-
-@login_required
-def cohort_edit(request, cohort_id):
-    school = request.user.profile.get_current_school(request)
-    if not (
-        request.user.profile.is_admin_for_school(school)
-        or request.user.profile.is_manager_for_school(school)
-    ):
-        return redirect("cohort_list")
-
-    cohort = get_object_or_404(Cohort, id=cohort_id, school=school)
-
-    if request.method == "POST":
-        form = CohortForm(request.POST, instance=cohort)
-        if form.is_valid():
-            form.save()
-            return redirect("cohort_list")
-    else:
-        form = CohortForm(instance=cohort)
-
-    return render(
-        request, "profile/cohorts/cohort_form.html", {"form": form, "is_editing": True}
-    )
-
-
-@login_required
-def cohort_delete(request, cohort_id):
-    school = request.user.profile.get_current_school(request)
-    if not (
-        request.user.profile.is_admin_for_school(school)
-        or request.user.profile.is_manager_for_school(school)
-    ):
-        return redirect("cohort_list")
-
-    cohort = get_object_or_404(Cohort, id=cohort_id, school=school)
-
-    if request.method == "POST":
-        password = request.POST.get("password")
-        user = authenticate(username=request.user.username, password=password)
-
-        if user is not None:
-            cohort.delete()
-            return redirect("cohort_list")
-
-        error_message = "Incorrect password. Please try again."
-        return render(
-            request,
-            "profile/cohorts/cohort_confirm_delete.html",
-            {"cohort": cohort, "error_message": error_message},
-        )
-
-    return render(request, "profile/cohorts/cohort_confirm_delete.html", {"cohort": cohort})
